+++ src/utils/configUtils.ts
@@ -25,6 +25,7 @@ import {
   isFolderOwned,
   takeOwnershipBufferDir,
   takeOwnershipStorageDir,
+  getDriveFormat,
 } from 'main/util';
 
 const allowRecordCategory = (cfg: ConfigService, category: VideoCategory) => {
@@ -261,6 +262,37 @@ const getLocaleError = (phrase: Phrase) => {
   return getLocalePhrase(lang, phrase);
 };
 
+const validateLogPathFilesystem = async (
+  logPath: string,
+  logPathLabel: Phrase,
+) => {
+  const driveFormat = await getDriveFormat(logPath);
+
+  if (!driveFormat) {
+    // The above can return undefined. If it does just log it and move
+    // on. We don't want to block the whole setup if for example powershell
+    // isn't available for some reason.
+    console.warn('[Util] Skipping NTFS filesystem check for', logPath);
+    return;
+  }
+
+  // NTFS is the only supported filesystem for reading the combat log due
+  // to our reliance on the Node watch API.
+  if (driveFormat.toLowerCase() !== 'ntfs') {
+    console.error(
+      '[Util] Unsupported filesystem for WoW log path',
+      driveFormat,
+      logPath,
+    );
+
+    const error = `${getLocaleError(logPathLabel)}: ${getLocaleError(
+      Phrase.UnsupportedWowFilesystem,
+    )}`;
+
+    throw new Error(error);
+  }
+};
+
 const validateBaseConfig = async (config: BaseConfig) => {
   const {
     storagePath,
@@ -359,6 +391,8 @@ const validateBaseConfig = async (config: BaseConfig) => {
       const error = getLocaleError(Phrase.InvalidRetailLogPath);
       throw new Error(error);
     }
+
+    await validateLogPathFilesystem(retailLogPath, Phrase.RetailLogPathLabel);
   }
 
   if (recordRetailPtr) {
@@ -376,6 +410,11 @@ const validateBaseConfig = async (config: BaseConfig) => {
       const error = getLocaleError(Phrase.InvalidRetailPtrLogPathText);
       throw new Error(error);
     }
+
+    await validateLogPathFilesystem(
+      retailPtrLogPath,
+      Phrase.RetailPtrLogPathLabel,
+    );
   }
 
   if (recordClassic) {
@@ -393,6 +432,8 @@ const validateBaseConfig = async (config: BaseConfig) => {
       const error = getLocaleError(Phrase.InvalidClassicLogPath);
       throw new Error(error);
     }
+
+    await validateLogPathFilesystem(classicLogPath, Phrase.ClassicLogPathLabel);
   }
 
   if (recordClassicPtr) {
@@ -409,6 +450,11 @@ const validateBaseConfig = async (config: BaseConfig) => {
       const error = getLocaleError(Phrase.InvalidClassicPtrLogPath);
       throw new Error(error);
     }
+
+    await validateLogPathFilesystem(
+      classicPtrLogPath,
+      Phrase.ClassicPtrLogPathLabel,
+    );
   }
 
   if (recordEra) {
@@ -426,6 +472,8 @@ const validateBaseConfig = async (config: BaseConfig) => {
       const error = getLocaleError(Phrase.InvalidEraLogPath);
       throw new Error(error);
     }
+
+    await validateLogPathFilesystem(eraLogPath, Phrase.ClassicEraLogPathLabel);
   }
 };